+++ src/methods/types/methodQueue.js
@@ -7,7 +7,7 @@ import Queue from '../../types/Queue';
  * @function method.queue
  *
  * @arg {Object}   [options]
- * @arg {Function} [options.set] - Called after a new callback is added to the queue. Provides a reference to the queue, sets the context to the methods constructor.
+ * @arg {Function} [options.set] - Called after a new callback is added to the queue. Provides a reference to the queue, the new ID for the callback, the callback, and sets the context to the methods constructor.
  *
  * @returns {Function} accepts a new value and returns the methods constructor (allows chaining), or if no args are passed returns the instance of Queue
  */
@@ -27,10 +27,10 @@ export default (options = {}) => {
 
 		if (arguments.length) {
 			if (isFunction(callback)) {
-				this[key].add(callback);
+				const ID = this[key].add(callback);
 
 				if (options.set) {
-					options.set.call(this, this[key]);
+					options.set.call(this, this[key], ID, callback);
 				}
 			}