+++ src/utility/difference.js
@@ -1,4 +1,0 @@
-export default (array1, ...args) => {
-	let diffArrays = [].concat(...args);
-	return array1.filter((item1) => diffArrays.every((item2) => item1 !== item2));
-};