+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.3.2] - 2019-2-14
+### Changed
+- Replaced dependencies on clone and deepEqual with object-agent
+- Added ID and callback args to the methodQueue set option
+
 ## [0.3.1] - 2019-2-10
 ### Added
 - [castArray](docs/castArray.md)
@@ -102,6 +107,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial enforcers, methods, and documentation
 
+[0.3.2]: https://github.com/DarrenPaulWright/type-enforcer/compare/v0.3.1...c0.3.2
 [0.3.1]: https://github.com/DarrenPaulWright/type-enforcer/compare/v0.3.0...c0.3.1
 [0.3.0]: https://github.com/DarrenPaulWright/type-enforcer/compare/v0.2.4...c0.3.0
 [0.2.4]: https://github.com/DarrenPaulWright/type-enforcer/compare/v0.2.3...c0.2.4